init navbar
